# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Petshion API",
    "version": "1.0.0",
    "description": "Petshion API with express"
  },
  "host": "petshion.herokuapp.com",
  "basePath": "/",
  "tags": [],
  "schemes": [
    "https",
    "http"
  ],
  "consumes": [],
  "produces": [],
  "paths": {
    "/product": {
      "post": {
        "tags": [
          "Product"
        ],
        "summary": "DB에 상품정보 등록",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "example": "any"
                },
                "images": {
                  "example": "any"
                },
                "Brand_name": {
                  "example": "any"
                },
                "price": {
                  "example": "any"
                },
                "kind": {
                  "example": "any"
                },
                "content": {
                  "example": "any"
                },
                "size": {
                  "example": "any"
                },
                "color": {
                  "example": "any"
                },
                "tag": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {}
      }
    },
    "/product/{product_id}": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "DB에 product_id의 상품정보 요청",
        "description": "",
        "parameters": [
          {
            "name": "product_id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "404": {
            "description": "찾을 수 없음"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/main": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "메인화면 요청",
        "description": "",
        "parameters": [],
        "responses": {
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/search": {
      "get": {
        "tags": [
          "Product"
        ],
        "summary": "필터링 검색 결과",
        "description": "",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "description": "종",
            "type": "array",
            "collectionFormat": "multi"
          },
          {
            "name": "size",
            "in": "query",
            "description": "크기",
            "type": "array",
            "collectionFormat": "multi"
          },
          {
            "name": "color",
            "in": "query",
            "description": "색깔",
            "type": "array",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "404": {
            "description": "찾을 수 없음"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/pawmark": {
      "get": {
        "tags": [
          "PawMark"
        ],
        "summary": "발도장 조회",
        "description": "",
        "parameters": [],
        "responses": {
          "401": {
            "description": "권한 없음"
          }
        }
      }
    },
    "/pawmark/add": {
      "post": {
        "tags": [
          "PawMark"
        ],
        "summary": "발도장 추가",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "pawmark": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "권한 없음"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/pawmark/remove": {
      "delete": {
        "tags": [
          "PawMark"
        ],
        "summary": "발도장 제거",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "pawmark": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "권한 없음"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/basket": {
      "get": {
        "tags": [
          "Basket"
        ],
        "summary": "장바구니 조회",
        "description": "",
        "parameters": [],
        "responses": {
          "401": {
            "description": "권한 없음"
          }
        }
      }
    },
    "/basket/add": {
      "post": {
        "tags": [
          "Basket"
        ],
        "summary": "장바구니 추가",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "product_id": {
                  "example": "any"
                },
                "selected_color": {
                  "example": "any"
                },
                "selected_size": {
                  "example": "any"
                },
                "selected_count": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "권한 없음"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    },
    "/basket/remove": {
      "delete": {
        "tags": [
          "Basket"
        ],
        "summary": "장바구니 제거",
        "description": "",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "401": {
            "description": "권한 없음"
          },
          "500": {
            "description": "내부 서버 오류"
          }
        }
      }
    }
  },
  "definitions": {}
}